+++ src/openforms/js/components/admin/form_design/registrations/camunda/index.js
@@ -1,0 +1,3 @@
+import CamundaOptionsForm from './CamundaOptionsForm';
+
+export default CamundaOptionsForm;
